# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "确定性八字排盘 API",
    "description": "立春换年、节令换月、真太阳时23点换日的可审计排盘服务。",
    "version": "0.3.0"
  },
  "paths": {
    "/api/v1/charts": {
      "post": {
        "summary": "Create Chart",
        "description": "执行完整排盘；路由层不直接参与历法计算。",
        "operationId": "create_chart_api_v1_charts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Chart Api V1 Charts Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses": {
      "post": {
        "summary": "Create Analysis",
        "description": "持久化命盘快照和八个待处理板块，不在API进程内调用模型。",
        "operationId": "create_analysis_api_v1_analyses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{job_id}/status": {
      "get": {
        "summary": "Analysis Status",
        "operationId": "analysis_status_api_v1_analyses__job_id__status_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{job_id}": {
      "get": {
        "summary": "Analysis Result",
        "operationId": "analysis_result_api_v1_analyses__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Analysis",
        "operationId": "delete_analysis_api_v1_analyses__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{job_id}/sections/{section_code}/retry": {
      "post": {
        "summary": "Retry Analysis Section",
        "operationId": "retry_analysis_section_api_v1_analyses__job_id__sections__section_code__retry_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "section_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Section Code"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{job_id}/cancel": {
      "post": {
        "summary": "Cancel Analysis",
        "operationId": "cancel_analysis_api_v1_analyses__job_id__cancel_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/charts/{chart_id}": {
      "delete": {
        "summary": "Delete Persisted Chart",
        "description": "删除出生信息、排盘快照以及其全部分析和调用日志。",
        "operationId": "delete_persisted_chart_api_v1_charts__chart_id__delete",
        "parameters": [
          {
            "name": "chart_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chart Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/wechat/start": {
      "get": {
        "summary": "Start Wechat Login",
        "operationId": "start_wechat_login_api_v1_auth_wechat_start_get",
        "parameters": [
          {
            "name": "return_to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 500,
              "default": "/",
              "title": "Return To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/wechat/callback": {
      "get": {
        "summary": "Finish Wechat Login",
        "operationId": "finish_wechat_login_api_v1_auth_wechat_callback_get",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 1024,
              "title": "Code"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 1024,
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/me/analyses": {
      "get": {
        "summary": "My Analyses",
        "operationId": "my_analyses_api_v1_me_analyses_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnalysisHistoryItem"
                  },
                  "title": "Response My Analyses Api V1 Me Analyses Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/locations/cities": {
      "get": {
        "summary": "Cities",
        "description": "按省级行政区名称返回可测算城市及其 location_id。",
        "operationId": "cities_api_v1_locations_cities_get",
        "parameters": [
          {
            "name": "province",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 20,
              "title": "Province"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Cities Api V1 Locations Cities Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalysisAccepted": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "chart_id": {
            "type": "string",
            "title": "Chart Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "partial",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "deduplicated": {
            "type": "boolean",
            "title": "Deduplicated",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "chart_id",
          "status"
        ],
        "title": "AnalysisAccepted"
      },
      "AnalysisHistoryItem": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "chart_id": {
            "type": "string",
            "title": "Chart Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "birth_local_datetime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Local Datetime"
          },
          "completed_sections": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Completed Sections"
          },
          "total_sections": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Sections"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "chart_id",
          "status",
          "completed_sections",
          "total_sections",
          "created_at"
        ],
        "title": "AnalysisHistoryItem"
      },
      "AnalysisResultResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "chart_id": {
            "type": "string",
            "title": "Chart Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "partial",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "prompt_version": {
            "type": "string",
            "title": "Prompt Version"
          },
          "request_hash": {
            "type": "string",
            "title": "Request Hash"
          },
          "cancel_requested": {
            "type": "boolean",
            "title": "Cancel Requested"
          },
          "completed_sections": {
            "type": "integer",
            "title": "Completed Sections"
          },
          "failed_sections": {
            "type": "integer",
            "title": "Failed Sections"
          },
          "total_sections": {
            "type": "integer",
            "title": "Total Sections"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "chart": {
            "additionalProperties": true,
            "type": "object",
            "title": "Chart"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/AnalysisSection"
            },
            "type": "array",
            "title": "Sections"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "chart_id",
          "status",
          "model_id",
          "prompt_version",
          "request_hash",
          "cancel_requested",
          "completed_sections",
          "failed_sections",
          "total_sections",
          "created_at",
          "chart",
          "sections",
          "disclaimer"
        ],
        "title": "AnalysisResultResponse"
      },
      "AnalysisSection": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "char_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Char Count"
          },
          "length_status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "ok",
                  "short",
                  "long"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Length Status"
          },
          "retry_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Retry Count"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "code",
          "title",
          "status",
          "char_count",
          "retry_count"
        ],
        "title": "AnalysisSection"
      },
      "AnalysisStatusResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "chart_id": {
            "type": "string",
            "title": "Chart Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "partial",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "prompt_version": {
            "type": "string",
            "title": "Prompt Version"
          },
          "request_hash": {
            "type": "string",
            "title": "Request Hash"
          },
          "cancel_requested": {
            "type": "boolean",
            "title": "Cancel Requested"
          },
          "completed_sections": {
            "type": "integer",
            "title": "Completed Sections"
          },
          "failed_sections": {
            "type": "integer",
            "title": "Failed Sections"
          },
          "total_sections": {
            "type": "integer",
            "title": "Total Sections"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "chart_id",
          "status",
          "model_id",
          "prompt_version",
          "request_hash",
          "cancel_requested",
          "completed_sections",
          "failed_sections",
          "total_sections",
          "created_at"
        ],
        "title": "AnalysisStatusResponse"
      },
...
      "ChartRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "birth_local_datetime": {
            "type": "string",
            "title": "Birth Local Datetime"
          },
          "location_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Location Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "gender",
          "birth_local_datetime",
          "location_id"
        ],
        "title": "ChartRequest",
        "description": "排盘输入模型；禁止额外字段，防止未支持的流派参数混入。"
      },
      "Gender": {
        "type": "string",
        "enum": [
          "male",
          "female"
        ],
        "title": "Gender",
        "description": "大运顺逆规则只接受男、女两个确定选项。"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}